Add log for UpdateModel & UserModel

+++ model/User_Model.php
@@ -10,6 +10,7 @@ class User_Model extends Model
 
         $this->loadHelper("randomizer");
         $this->loadModel("perms");
+        $this->loadModel("log");
         $this->loadHelper("db");
         $this->loadHelper("session");
         $this->loadHelper("url");
@@ -17,6 +18,14 @@ class User_Model extends Model
 
     }
 
+    /**
+     * @description Create our first User for software
+     * @used Only on installation
+     * @param $userName
+     * @param $password
+     * @return bool
+     * @see Install_Model::makeInstall()
+     */
     public function createAdminUser($userName, $password)
     {
         $userName = htmlspecialchars(trim($userName));
@@ -36,6 +45,9 @@ class User_Model extends Model
         return $this->helper->db->insert($this->table, $data) != false;
     }
 
+    /**
+     * @param $request
+     */
     public function addUser($request)
     {
 
@@ -85,13 +97,20 @@ class User_Model extends Model
                 'hasCaisse' => $caisse,
                 'isDefaultPass' => $defaultPass
             );
-            if (!$this->helper->db->insert($this->table, $data)) {
-                $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors de l'ajout de l'utilisater");
-                $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+
+            if ($this->model->log->addLog(wp_get_current_user()->user_login . "(site)", "UserModel", "Create", "", "", json_encode($data))) {
+                if (!$this->helper->db->insert($this->table, $data)) {
+                    $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors de l'ajout de l'utilisateur");
+                    $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+                } else {
+                    $this->helper->session->set_flashdata("success", "L'utilisateur' a bien été ajouté");
+                    $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+                }
             } else {
-                $this->helper->session->set_flashdata("success", "L'utilisateur' a bien été ajouté");
+                $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors de l'ajout du log. L'utilisateur n'as pas été ajouté");
                 $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
             }
+
         }
 
     }
@@ -111,6 +130,9 @@ class User_Model extends Model
         return $this->helper->db->get_like($this->table, array('usernameShopLogin' => array($request->get('searchUserName'), 'before')));
     }
 
+    /**
+     * @param $request
+     */
     public function editUser($request)
     {
         $id = $request->get('id');
@@ -140,14 +162,18 @@ class User_Model extends Model
                 'hasStock' => $stock,
                 'hasCaisse' => $caisse,
             );
-            if (!$this->helper->db->update($this->table, $data, array('idShopLogin' => $id))) {
-                $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors de la modification de l'utilisater");
-                $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+            if ($this->model->log->addLog(wp_get_current_user()->user_login . "(site)", "UserModel", "Edit", $id, json_encode($oldUser), json_encode($data))) {
+                if (!$this->helper->db->update($this->table, $data, array('idShopLogin' => $id))) {
+                    $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors de la modification de l'utilisateur");
+                    $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+                } else {
+                    $this->helper->session->set_flashdata("success", "L'utilisateur' a bien été modifié");
+                    $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+                }
             } else {
-                $this->helper->session->set_flashdata("success", "L'utilisateur' a bien été modifié");
+                $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors l'ajout du log. L'utilisateur n'as pas été modifié");
                 $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
             }
-
         } else {
             $this->helper->session->set_flashdata("error", "L'utilisateur spécifié est introuvable");
             $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
@@ -170,34 +196,54 @@ class User_Model extends Model
         return password_hash($password, PASSWORD_DEFAULT);
     }
 
+    /**
+     * @param $request
+     */
     public function deleteUser($request)
     {
         $id = $request->get('userId');
         if ($this->isUserId($id)) {
-            if (!$this->helper->db->delete($this->table, array('idShopLogin' => $id))) {
-                $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors de la suppression de l'utilisater");
-                $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+            $oldUser = $this->getUserById($id);
+
+            if ($this->model->log->addLog(wp_get_current_user()->user_login . "(site)", "USerModel", "Delete", $id, json_encode($oldUser))) {
+                if (!$this->helper->db->delete($this->table, array('idShopLogin' => $id))) {
+                    $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors de la suppression de l'utilisateur");
+                    $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+                } else {
+                    $this->helper->session->set_flashdata("success", "L'utilisateur' a bien été supprimé");
+                    $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+                }
             } else {
-                $this->helper->session->set_flashdata("success", "L'utilisateur' a bien été supprimé");
+                $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors l'ajout du log. L'utilisateur n'as pas été supprimé");
                 $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
             }
+
         } else {
             $this->helper->session->set_flashdata("error", "L'utilisateur spécifié est introuvable");
             $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
         }
     }
 
+    /**
+     * @param $request
+     */
     public function resetPassword($request)
     {
         $id = $request->get('userId');
         if ($this->isUserId($id)) {
             $data = array('passwordShopLogin' => $this->helper->randomizer->randomizeString(8),
                 'isDefaultPass' => 1);
-            if (!$this->helper->db->update($this->table, $data, array('idShopLogin' => $id))) {
-                $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors de la reinitialisation du mot de passe de l'utilisater");
-                $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+
+            if ($this->model->log->addLog(wp_get_current_user()->user_login . "(site)", "USerModel", "ResetPass", $id)) {
+                if (!$this->helper->db->update($this->table, $data, array('idShopLogin' => $id))) {
+                    $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors de la reinitialisation du mot de passe de l'utilisateur");
+                    $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+                } else {
+                    $this->helper->session->set_flashdata("success", "Le mot de passe utilisateur' a bien été réinitialisé");
+                    $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
+                }
             } else {
-                $this->helper->session->set_flashdata("success", "Le mot de passe utilisateur' a bien été réinitialisé");
+                $this->helper->session->set_flashdata("error", "Une erreur interne est survenue lors de l'ajout du log. La reinitialisation du mot de passe de l'utilisateur n'as pas été effectuée");
                 $this->helper->url->redirect("wp-admin/admin.php?page=TijaraShop/users");
             }
         } else {
@@ -206,19 +252,31 @@ class User_Model extends Model
         }
     }
 
-    public function updatePassword($idShopLogin, $newPass)
+    /**
+     * @param $idShopLogin
+     * @param $newPass
+     * @return stdClass
+     */
+    public function updatePassword($idShopLogin, $newPass, $userName)
     {
         $data = array(
             "passwordShopLogin" => $newPass,
             "isDefaultPass" => 0
         );
         $obj = new stdClass();
-        if (!$this->helper->db->update($this->table, $data, array('idShopLogin' => $idShopLogin))) {
-            $obj->state = 0;
-            $obj->error = "Erreur interne dans l'api lors de la mise a jour du mot de passe";
+        
+        if ($this->model->log->addLog($userName, "UserModel", "ChangePass", $idShopLogin)) {
+            if (!$this->helper->db->update($this->table, $data, array('idShopLogin' => $idShopLogin))) {
+                $obj->state = 0;
+                $obj->error = "Erreur interne dans l'api lors de la mise a jour du mot de passe";
+            } else {
+                $obj->state = 1;
+            }
         } else {
-            $obj->state = 1;
+            $obj->state = 0;
+            $obj->error = "Erreur interne lors de l'ajout du log. Le mot de passe n'as pas été modifié";
         }
+
         return $obj;
     }
